refactor: 파일 구조 및 위치 변경
