+++ src/main/java/com/kakaotechcampus/journey_planner/application/waypoint/WaypointService.java
@@ -1,11 +1,11 @@
-package com.kakaotechcampus.journey_planner.application;
+package com.kakaotechcampus.journey_planner.application.waypoint;
 
 import com.kakaotechcampus.journey_planner.domain.plan.Plan;
 import com.kakaotechcampus.journey_planner.domain.plan.PlanRepository;
 import com.kakaotechcampus.journey_planner.domain.waypoint.Location;
 import com.kakaotechcampus.journey_planner.domain.waypoint.Waypoint;
 import com.kakaotechcampus.journey_planner.domain.waypoint.WaypointRepository;
-import com.kakaotechcampus.journey_planner.presentation.dto.WaypointRequest;
+import com.kakaotechcampus.journey_planner.presentation.dto.waypoint.WaypointRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
