+++ src/main/java/com/kakaotechcampus/journey_planner/presentation/waypoint/WaypointController.java
@@ -1,9 +1,9 @@
-package com.kakaotechcampus.journey_planner.presentation;
+package com.kakaotechcampus.journey_planner.presentation.waypoint;
 
-import com.kakaotechcampus.journey_planner.application.WaypointService;
+import com.kakaotechcampus.journey_planner.application.waypoint.WaypointService;
 import com.kakaotechcampus.journey_planner.domain.waypoint.Waypoint;
 import com.kakaotechcampus.journey_planner.domain.waypoint.WaypointMapper;
-import com.kakaotechcampus.journey_planner.presentation.dto.WaypointRequest;
+import com.kakaotechcampus.journey_planner.presentation.dto.waypoint.WaypointRequest;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
